Moving the getProductsByBillingSlug selector from data-store to client state
