+++ packages/data-stores/src/products-list/test/selectors.ts
@@ -45,26 +45,6 @@ describe( 'selectors', () => {
 				price_tiers: [],
 				price_tier_slug: '',
 			},
-			wp_mp_theme_tsubaki_test_yearly: {
-				product_id: 3001,
-				product_name: 'Tsubaki Third-Party Test',
-				product_slug: 'wp_mp_theme_tsubaki_test_yearly',
-				description: '',
-				product_type: 'marketplace_theme',
-				available: true,
-				billing_product_slug: 'wp-mp-theme-tsubaki-test',
-				is_domain_registration: false,
-				cost_display: 'US$100.00',
-				combined_cost_display: 'US$100.00',
-				cost: 100,
-				cost_smallest_unit: 10000,
-				currency_code: 'USD',
-				price_tier_list: [],
-				price_tier_usage_quantity: null,
-				product_term: 'year',
-				price_tiers: [],
-				price_tier_slug: '',
-			},
 		};
 
 		( wpcomRequest as jest.Mock ).mockResolvedValue( apiResponse );
@@ -85,8 +65,5 @@ describe( 'selectors', () => {
 
 		expect( select( store ).getProductsList() ).toEqual( apiResponse );
 		expect( select( store ).getProductBySlug( 'free_plan' ) ).toEqual( apiResponse[ 'free_plan' ] );
-		expect( select( store ).getProductsByBillingSlug( 'wp-mp-theme-tsubaki-test' ) ).toEqual( [
-			apiResponse[ 'wp_mp_theme_tsubaki_test_yearly' ],
-		] );
 	} );
 } );
