+++ packages/data-stores/src/products-list/types.ts
@@ -7,7 +7,6 @@ export interface Dispatch {
 
 export interface ProductsListItem {
 	available: boolean;
-	billing_product_slug: string;
 	combined_cost_display: string;
 	cost: number;
 	currency_code: string;
